// src/app/models/context.ts
export type Context = {
    solarDay: number;
    solarMonth: number;
    solarYear: number;
    lunarDay: number;
    lunarMonth: number;
    lunarYear: number;
    startLunarYear: number;
    endLunarYear: number;
    eventIndex: number;
    totalEvents: number;
    isFirstEvent: boolean;
    isLastEvent: boolean;
}